+++ CHANGELOG.rst
@@ -3,6 +3,7 @@ Changelog
 =========
 
 * Implementation of a new `singlechain` function. Goal: provide a function to return GE values directly from a PDB file
+* Remove `qentangled` module because it is not used anymore
 
 v0.1.0 (2022-12-15)
 ------------------------------------------------------------