+++ CHANGELOG.rst
@@ -2,6 +2,9 @@
 Changelog
 =========
 
+v0.1.1 (2023-03-20)
+------------------------------------------------------------
+
 * Implementation of a new `singlechain` function. Goal: provide a function to return GE values directly from a PDB file
 * Remove `qentangled` module because it is not used anymore